pyproject.toml
[project]
name = "latent-resonance"
version = "0.1.0"
description = "A computational instrument that transmutes facial micro-movements into generated audio spectrograms"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "librosa>=0.11.0",
    "mediapipe>=0.10.32",
    "numpy>=2.3.5",
    "opencv-python>=4.13.0.90",
    "pillow>=11.0.0",
    "soundfile>=0.13.1",
    "requests>=2.32.0",
    "python-dotenv>=1.2.1",
    "matplotlib>=3.9.0",
    "sounddevice>=0.5.0",
    "torchvision>=0.25.0",
    "torchaudio>=2.10.0",
    "torch>=2.10.0",
]

[project.scripts]
latent-resonance-dataset = "latent_resonance.dataset.cli:main"
latent-resonance-scraper = "latent_resonance.dataset.scraper:main"

[dependency-groups]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
